+++ src/components/AnimateSharedLayout/utils/crossfader.ts
@@ -1,4 +1,4 @@
-import { getFrameData } from "framesync"
+import sync, { getFrameData } from "framesync"
 import {
     circOut,
     linear,
@@ -66,20 +66,41 @@ export function createCrossfader(): Crossfader {
         const { lead, follow } = options
         isActive = true
         finalCrossfadeFrame = null
+        let hasUpdated = false
+
+        const onUpdate = () => {
+            console.log("crossfade update", lead?.current.id, progress.get())
+            hasUpdated = true
+            lead && lead.scheduleRender()
+            follow && follow.scheduleRender()
+        }
+
+        const onComplete = () => {
+            isActive = false
+
+            /**
+             * If the crossfade animation is no longer active, flag a frame
+             * that we're still allowed to crossfade
+             */
+            finalCrossfadeFrame = getFrameData().timestamp
+        }
+
         return animate(progress, target, {
             ...transition,
-            onUpdate: () => {
-                lead && lead.scheduleRender()
-                follow && follow.scheduleRender()
-            },
+            onUpdate,
             onComplete: () => {
-                isActive = false
+                if (!hasUpdated) {
+                    progress.set(1)
+                    /**
+                     * If we never ran an update, for instance if this was an instant transition, fire a
+                     * simulated final frame to ensure the crossfade gets applied correctly.
+                     */
+                    sync.read(onComplete)
+                } else {
+                    onComplete()
+                }
 
-                /**
-                 * If the crossfade animation is no longer active, flag a frame
-                 * that we're still allowed to crossfade
-                 */
-                finalCrossfadeFrame = getFrameData().timestamp
+                onUpdate()
             },
         } as any)
     }
@@ -114,6 +135,7 @@ export function createCrossfader(): Crossfader {
          */
         const leadTargetOpacity = (latestLeadValues.opacity as number) ?? 1
         const followTargetOpacity = (latestFollowValues?.opacity as number) ?? 1
+
         if (options.crossfadeOpacity && follow) {
             leadState.opacity = mix(0, leadTargetOpacity, easeCrossfadeIn(p))
             followState.opacity = options.preserveFollowOpacity
@@ -134,9 +156,12 @@ export function createCrossfader(): Crossfader {
     }
 
     return {
-        isActive: () =>
-            leadState &&
-            (isActive || getFrameData().timestamp === finalCrossfadeFrame),
+        isActive: () => {
+            return (
+                leadState &&
+                (isActive || getFrameData().timestamp === finalCrossfadeFrame)
+            )
+        },
         fromLead(transition) {
             return startCrossfadeAnimation(0, transition)
         },