+++ src/render/types.ts
@@ -13,6 +13,7 @@ import { MotionValue } from "../value"
 import { AnimationState } from "./utils/animation-state"
 import { LifecycleManager } from "./utils/lifecycles"
 import { LayoutState, TargetProjection } from "./utils/state"
+import { LayoutTree } from "./utils/tree"
 
 export interface MotionPoint {
     x: MotionValue<number>
@@ -24,6 +25,7 @@ export interface VisualElement<Instance = any, RenderState = any>
     treeType: string
     depth: number
     current: Instance | null
+    layoutTree: LayoutTree
     manuallyAnimateOnMount: boolean
     blockInitialAnimation?: boolean
     presenceId: number | undefined
@@ -35,7 +37,6 @@ export interface VisualElement<Instance = any, RenderState = any>
     isStatic?: boolean
     getInstance(): Instance | null
     path: VisualElement[]
-    addChild(child: VisualElement): () => void
     sortNodePosition(element: VisualElement): number
 
     addVariantChild(child: VisualElement): undefined | (() => void)
@@ -111,6 +112,7 @@ export interface VisualElement<Instance = any, RenderState = any>
     stopLayoutAnimation(): void
     snapshotViewportBox(): void
     updateLayoutProjection(): void
+    updateTreeLayoutProjection(): void
     makeTargetAnimatable(
         target: TargetAndTransition,
         isLive?: boolean