+++ src/render/utils/sort-by-depth.ts
@@ -1,0 +1,4 @@
+import { VisualElement } from "../types"
+
+export const sortByDepth = (a: VisualElement, b: VisualElement) =>
+    a.depth - b.depth